Amicable Nums
I have finally finished the Nth amicable number finder. It's not well optimized, so you will have to wait a long time for even the most basic of amicable numbers. Keep in mind I'm not using any mathematical formula to make this, since I know one does exist. Just run the program, with the amicable number you want to find as one of the arguments. It also has functionality that may be able to find connections between three or more numbers, but as of now I think it may be impossible. Expect performance upgrades soon, and perhaps a solution to the unsolvable problem.