Convert to Base 2 - Electric Boogalo
It's the same program and it works in the same way, except now it can take in and output numbers all the way up to base 36. Anything beyond that would require new symbols be made, so it shouldn't get any higher than that. Functionality is still the same. Do this:

python3 ConvertToBase.py [Starting number] [Starting base] [Ending base]

Note that the digits 11-35 are the letters a-z, like in hexadecimal.