Even Better Sudoku Solver
Alright, I've ironed out all the kinks, and now it works 100% of the time, provided it's given a puzzle with a solution. I was experimenting if BFS of DFS was faster, but after many experiments there didn't seem to be that drastic of a difference. It works the same as before, but now it's really good.