Tic Tac Toe Solver V2
All there is so far is a function that finds all boards, games, wins for x and o, and ties given a board. Starting with a blank board, there are 5478 possible boards that can be accessed, and there are 255168 possible games, with 49392 wins for x and 77904 wins for o, along with 127872 ties. Just goes to show how stupid tic tac toe is; about half the possible games just end with a tie.