Connect Four V2.0
The first version of online connect four using Java is done! You can play against someone online, or you can be an AI and play against yourself. There are some bugs to work out, and some things could be fixed, but other than that I'd say go ahead and try it. Granted, I've only tried it on my own computer, so I'm not completely sure that the networking stuff works, but I think it does.