Neural Net V0.6
Still working on it. I think I made a very small error that's screwing the entire program over, but I just haven't found it yet. My network is fairly optimized, but it caps out at around an error of 0.19 or so. Maybe that is the best it can do, or perhaps the error mentioned above may be causing it.